Add files via upload
added the simclr test files I've developed the last week as well as an attempt at my data_utils. 

files:
grab_data - this is to move 20k files for the unlabeled dataset for quicker training
simclr model V2 - the initial test, its V2 because its not the keras.io article on a different dataset. Changes from the keras.io article are largely the inclusion of the Resnet50 as opposed to the CNN encoder
simclr model V2 100 epocs - same structure as V2, only 100 epochs instead of 20 (oof...I misspelled epochs...)
data_utils - changes to how the datasets are prepared as we're not using a tfds structure